kubernetes: option to limit namespaces when listing pods
This adds an environment variable that allows for limiting the pods that
sky lists in the cluster.

It can be useful in cases when running in a cluster with thousands of
pods where the client knows that only pods in a subset of namespaces
will actually be using GPU resources.